rename kontena-lib to kontena-common, using conventional common/lib/... layout
