+++ common/kontena-common.gemspec
@@ -1,12 +1,12 @@
 # TODO: version from ./VERSION
 
 Gem::Specification.new do |s|
-  s.name          = 'kontena-lib'
+  s.name          = 'kontena-common'
   s.version       = '1.4.0.dev'
-  s.summary       = "Kontena libraries"
+  s.summary       = "Common Kontena libraries"
   s.authors       = ["Kontena, Inc"]
   s.email         = ["info@kontena.io"]
-  s.description   = "Common code shared by Kontena components"
+  s.description   = "Common libraries shared by Kontena components"
   s.homepage      = "https://www.kontena.io"
   s.license       = "Apache-2.0"
