+++ common/lib/kontena-common.rb
