+++ common/lib/kontena-wait.rb
