+++ common/lib/kontena/logging.rb
