+++ common/lib/kontena/wait/helper.rb
