+++ common/spec/spec_helper.rb
@@ -13,7 +13,7 @@
 # it.
 
 
-require 'kontena-lib'
+require 'kontena-common'
 
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
