--- lib/.rspec
@@ -1,1 +1,0 @@
--I . --require spec_helper
